+++ windows/ingredientsWindow/check_iWindow.java
@@ -51,7 +51,7 @@ public class check_iWindow extends abstractEdit_CheckWindow {
             String id = Integer.toString(temp.getId());
             String prov_id = Integer.toString(temp.getProviderID());
             String price = Float.toString(temp.getPrice());
-            String amount = Integer.toString(temp.getAmount());
+            String amount = Float.toString(temp.getAmount());
             String in_inventory;
             String active;
             if (temp.getInInventory())
