+++ windows/menuWindow/assist_assist_edit_mWindow.java
@@ -93,10 +93,10 @@ public class assist_assist_edit_mWindow extends abstractAddWindow {
 
                 if (changeProductsButton.getText().equals("Don't Change Products Too")) {
                     Stack<Integer> stackIDs = new Stack<Integer>();
-                    Stack<Integer> stackQtys = new Stack<Integer>();
+                    Stack<Float> stackQtys = new Stack<Float>();
                     for (int i = 0; i < modelSelected.getRowCount(); i++) {
                         stackIDs.push(Integer.parseInt((String) modelSelected.getValueAt(i, 0)));
-                        stackQtys.push(Integer.parseInt((String) modelSelected.getValueAt(i, 4)));
+                        stackQtys.push(Float.parseFloat((String) modelSelected.getValueAt(i, 4)));
                     }
                     theManagerDB.updateProducts(theCurrentMenu.getId(), stackIDs, stackQtys);
                 }
@@ -208,7 +208,7 @@ public class assist_assist_edit_mWindow extends abstractAddWindow {
             String amountUsed = "Type here";
             if (alreadySelected) {
                 int tempAmount = new productAPI().getAmountOfProductInMenu(theCurrentMenu.getId(), temp.getId());
-                amountUsed = Integer.toString(tempAmount);
+                amountUsed = Float.toString(tempAmount);
             }
             tempModel.addRow(new String[] { productID, date, name, price, amountUsed, "Yes" });
         }
